feat(useResizableContent): update logic
