+++ src/hooks/useResizableContent/__stand__/useResizableContent.stand.mdx
@@ -4,7 +4,7 @@ import { UseResizableContentBasic } from './examples/UseResizableContentBasic/Us
 import { useResizableContent } from '@consta/uikit/useResizableContent';
 ```
 
-Хук `useResizableContent` позволяет сделать блоки в контейнере с изменяемыми размерами.
+С помощью хука `useResizableContent` размеры блоков в контейнере можно сделать изменяемыми.
 
 ```tsx
 export const UseResizableContentBasic = () => {
